+++ Write_Sort_Files.cpp
@@ -19,7 +19,7 @@ int Write_Sort_Files (const char* filename_write , struct Inf_Lines* arr_structs
 
     }
 
-    for (size_t i = 0 ; i < Onegin->num_lines ; i++) {
+    for (int i = 0 ; i < Onegin->num_lines ; i++) {
 
         fputs (arr_structs[i].str , fh);
         fputc ('\n' , fh);
